修复弓箭物品结算显示"未知"、消除18个编译警告、窗口最大化自适应
- cnName: 3个重复lambda提取为getMaterialCNName共享函数，补全弓/铁箭/蛛丝箭映射
- Camera: 新增updateProjection方法，窗口大小变化时重建投影矩阵
- RenderMaster: 每帧更新glViewport适配窗口尺寸
- Application: 捕获sf::Event::Resized转发给Camera
- 清理: 18个C4244/C4100/C4267/C4456编译警告

+++ MineCraft-One-Week-Challenge/Source/Maths/Vector2XZ.h
@@ -19,7 +19,7 @@ template <> struct hash<VectorXZ> {
         auto hash1 = hasher(vect.x);
         auto hash2 = hasher(vect.z);
 
-        return std::hash<decltype(vect.x)>{}(size_t((hash1 ^ hash2) >> 2));
+        return std::hash<decltype(vect.x)>{}(static_cast<int>((hash1 ^ hash2) >> 2));
     }
 };
 } // namespace std
@@ -35,7 +35,7 @@ template <> struct hash<sf::Vector3i> {
         auto hash3 = hasher(vect.z);
 
         return std::hash<decltype(vect.x)>{}(
-            size_t(hash1 ^ (hash2 << hash3) ^ hash3));
+            static_cast<int>(hash1 ^ (hash2 << hash3) ^ hash3));
     }
 };
 } // namespace std
