+++ MineCraft-One-Week-Challenge/Source/World/Chunk/Chunk.cpp
@@ -135,7 +135,7 @@ void Chunk::deleteMeshes()
 
 void Chunk::addSection()
 {
-    int y = m_chunks.size();
+    int y = static_cast<int>(m_chunks.size());
     m_chunks.emplace_back(sf::Vector3i(m_location.x, y, m_location.y),
                           *m_pWorld);
 }
